# apps/web/src/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "AccountPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AccountRecord"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "AccountPage",
        "type": "object"
      },
...
      "EventPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EventRecord"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "EventPage",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NarrativePage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/NarrativeRecord"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "items",
          "page"
        ],
        "title": "NarrativePage",
        "type": "object"
      },
...
      "QualityState": {
        "enum": [
          "LIVE",
          "STALE",
          "DEGRADED",
          "DISCONNECTED",
          "ERROR",
          "ESTIMATED"
        ],
        "title": "QualityState",
        "type": "string"
      },
...
    }
  },
  "info": {
    "contact": {
      "name": "Local operator"
    },
    "description": "Versioned Read Model API. It exposes no trading writes, no credential input, and no LIVE_TRADING unlock capability.",
    "license": {
      "name": "Private local project"
    },
    "summary": "Loopback-only, read-only P14 dashboard API",
    "title": "AegisQuant Read API",
    "version": "3.1.0.dev0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/accounts": {
      "get": {
        "operationId": "accounts_api_v1_accounts_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "quality_state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/QualityState"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Quality State"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Accounts",
        "tags": [
          "account"
        ]
      }
    },
    "/api/v1/accounts/{account_id}": {
      "get": {
        "operationId": "account_api_v1_accounts__account_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Account",
        "tags": [
          "account"
        ]
      }
    },
    "/api/v1/accounts/{account_id}/pnl": {
      "get": {
        "operationId": "account_pnl_api_v1_accounts__account_id__pnl_get",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PnLRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Account Pnl",
        "tags": [
          "account"
        ]
      }
    },
...
    "/api/v1/intelligence/events": {
      "get": {
        "operationId": "events_api_v1_intelligence_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "quality_state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/QualityState"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Quality State"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Events",
        "tags": [
          "intelligence"
        ]
      }
    },
    "/api/v1/intelligence/events/{event_cluster_id}": {
      "get": {
        "operationId": "event_api_v1_intelligence_events__event_cluster_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "event_cluster_id",
            "required": true,
            "schema": {
              "title": "Event Cluster Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Event",
        "tags": [
          "intelligence"
        ]
      }
    },
    "/api/v1/intelligence/events/{event_cluster_id}/evidence": {
      "get": {
        "operationId": "event_evidence_api_v1_intelligence_events__event_cluster_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "event_cluster_id",
            "required": true,
            "schema": {
              "title": "Event Cluster Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ClaimRecord"
                  },
                  "title": "Response Event Evidence Api V1 Intelligence Events  Event Cluster Id  Evidence Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Event Evidence",
        "tags": [
          "intelligence"
        ]
      }
    },
    "/api/v1/intelligence/narratives": {
      "get": {
        "operationId": "narratives_api_v1_intelligence_narratives_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "quality_state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/QualityState"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Quality State"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NarrativePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Narratives",
        "tags": [
          "intelligence"
        ]
      }
    },
...
    "/api/v1/intelligence/sources": {
      "get": {
        "operationId": "sources_api_v1_intelligence_sources_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "quality_state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/QualityState"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Quality State"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourcePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sources",
        "tags": [
          "intelligence"
        ]
      }
    },
    "/api/v1/models": {
      "get": {
        "operationId": "models_api_v1_models_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "quality_state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/QualityState"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Quality State"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Models",
        "tags": [
          "research"
        ]
      }
    },
...
    "/api/v1/risk/summary": {
      "get": {
        "operationId": "risk_summary_api_v1_risk_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "account_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RiskRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Risk Summary",
        "tags": [
          "risk"
        ]
      }
    },
...
  }
}